パーティクルシステムの初期化と描画処理を追加
GamePlayScene.cppにパーティクルエミッターの初期化と更新処理を追加しました。
GameEngine.cppではパーティクルの描画処理を新たに実装し、古い描画処理を削除しました。
ParticleManager.cppのパーティクル生成方法を変更し、MakeNewParticleメソッドを削除してParticleFactory::Createを使用するようにしました。
新たにParticleFactoryクラスを追加し、パーティクル生成ロジックを移動しました。
プロジェクトファイルのパスを修正し、ParticleEffectType.hを追加してエフェクトタイプを定義しました。
