+++ Project/EngineLayer/ParticleManagement/ParticleFactory.h
@@ -1,0 +1,17 @@
+#pragma once
+#include <random>
+#include "Vector3.h"
+#include "Particle.h"
+#include "ParticleEffectType.h"
+
+/// -------------------------------------------------------------
+///						パーティクル生成クラス
+/// -------------------------------------------------------------
+class ParticleFactory
+{
+public: /// ---------- メンバ関数 ---------- ///
+
+	// パーティクルを生成する関数
+	static Particle Create(std::mt19937& randomEngine, const Vector3& position, ParticleEffectType effectType);
+};
+
