+++ Project/ParticleEffectType.h
@@ -1,0 +1,9 @@
+#pragma once
+// エフェクトの種類を列挙型で定義
+enum class ParticleEffectType
+{
+	Default, 	// デフォルト
+	Slash,		// スラッシュ
+	Ring,		// リング
+	Cylinder,	// シリンダー
+};
