+++ Project/imgui.ini
@@ -28,7 +28,6 @@ Size=275,123
 [Window][Post Effect Settings]
 Pos=6,57
 Size=274,369
-Collapsed=1
 
 [Window][Collider]
 Pos=9,72
